EventComments entity creation, CRUD manager implementation

+++ RobinsonSportApp.Core/Managers/Associations/AssociationManager.cs
@@ -3,20 +3,14 @@ using Microsoft.EntityFrameworkCore;
 using RobinsonSportApp.Core.Constants;
 using RobinsonSportApp.Core.CustomExceptions;
 using RobinsonSportApp.Core.Managers.Associations.Models;
-using RobinsonSportApp.Core.Managers.Events.Models;
 using RobinsonSportApp.Data;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace RobinsonSportApp.Core.Managers.Associations
 {
     internal class AssociationManager(RobinsonSportAppDbContext _dbContext, IMapper _mapper) : IAssociationManager
     {
-        public async Task DeleteAssociationAsync(long id, CancellationToken cancellationToken = default)
+        public async Task DeleteAssociationAsync(int id, CancellationToken cancellationToken = default)
         {
             var association = await _dbContext.Associations.FirstOrDefaultAsync(e => e.Id == id, cancellationToken)
                                    ?? throw new RBException(ErrorMessages.AssociationNotFound, HttpStatusCode.NotFound);
